Class variable name is defined in parent namespace
Example:

class Foo:
  __namespace__ = "foo"
  fooStr = String

class Bar:
  __namespace__ = "bar"
  barStr = String
  var = Foo

In this example, the patch changes the namespace of "var" from "{foo}var" to "{bar}var".
This is consistent with the other member, "{bar}barStr". The "foo" namespace is only used
for "{foo}fooStr", and the name of the Foo class itself "{foo}Foo".

Also, if a method returns Bar ("_return=Bar"), the out variable name is defined in the
namespace of the service, instead of the Bar class:

class Service:
  @rpc(_return=String, _out_veriable_name="ret")
  def foo():
    pass

  @rpc(_return=Bar, _out_variable_name="ret")
  def bar():
    pass

In this example, the return variable of both functions are "{service}ret", instead of
"foo" returning "{service}ret", and "bar" returning "{bar}ret".