The relation between the Interface definition and its serialization as an interface document is changed from an is-a relation to a has-a relation.
The InterfaceBase class was split in two: Interface and InterfaceDocumentBase.

The Interface class creates and manages the abstract information to build the interface document.

The InterfaceDocumentBase class simply defines an interface for interface document implementors to implement. It generates the interface document using the information in the application's Interface instance.